# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: La Liga Brasil API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /teams/{slug}:
    get:
      operationId: getTeam
      tags: [teams]
      summary: Get team by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamWithArticles"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /admin/teams:
    get:
      operationId: adminListTeams
      tags: [admin]
      summary: List teams (admin)
      responses:
        "200":
          description: List of teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
    post:
      operationId: createTeam
      tags: [admin]
      summary: Create team
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamRequest"
      responses:
        "201":
          description: Created team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /admin/teams/{id}:
    put:
      operationId: updateTeam
      tags: [admin]
      summary: Update team
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamRequest"
      responses:
        "200":
          description: Updated team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
...
  /scraper/fetch:
    post:
      operationId: fetchFromSource
      tags: [scraper]
      summary: Fetch and translate articles from a source
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FetchSourceRequest"
      responses:
        "200":
          description: Fetched articles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FetchSourceResponse"
  /scraper/translate:
    post:
      operationId: translateArticle
      tags: [scraper]
      summary: Translate article content using AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranslateRequest"
      responses:
        "200":
          description: Translated content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranslateResponse"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Article:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        coverImage:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, published, scheduled]
        featured:
          type: boolean
        breakingNews:
          type: boolean
        category:
          type: string
        authorName:
          type: string
        teamId:
          type: integer
          nullable: true
        teamName:
          type: string
          nullable: true
        teamSlug:
          type: string
          nullable: true
        sourceUrl:
          type: string
          nullable: true
        sourceName:
          type: string
          nullable: true
        publishedAt:
          type: string
          nullable: true
        scheduledAt:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
        viewCount:
          type: integer
      required:
        - id
        - title
        - slug
        - excerpt
        - content
        - status
        - featured
        - breakingNews
        - category
        - authorName
        - createdAt
        - updatedAt
        - viewCount
...
    Team:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        shortName:
          type: string
        city:
          type: string
        stadium:
          type: string
        foundedYear:
          type: integer
          nullable: true
        description:
          type: string
          nullable: true
        logoUrl:
          type: string
          nullable: true
        primaryColor:
          type: string
        secondaryColor:
          type: string
        articleCount:
          type: integer
      required:
        - id
        - name
        - slug
        - shortName
        - city
        - stadium
        - primaryColor
        - secondaryColor
        - articleCount
    TeamWithArticles:
      allOf:
        - $ref: "#/components/schemas/Team"
        - type: object
          properties:
            recentArticles:
              type: array
              items:
                $ref: "#/components/schemas/Article"
          required:
            - recentArticles
...
    CreateTeamRequest:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        shortName:
          type: string
        city:
          type: string
        stadium:
          type: string
        foundedYear:
          type: integer
          nullable: true
        description:
          type: string
          nullable: true
        logoUrl:
          type: string
          nullable: true
        primaryColor:
          type: string
        secondaryColor:
          type: string
      required:
        - name
        - slug
        - shortName
        - city
        - stadium
        - primaryColor
        - secondaryColor
...
    FetchSourceRequest:
      type: object
      properties:
        sourceId:
          type: string
        maxArticles:
          type: integer
          default: 10
        autoImport:
          type: boolean
          default: false
      required:
        - sourceId
    FetchSourceResponse:
      type: object
      properties:
        articles:
          type: array
          items:
            $ref: "#/components/schemas/ScrapedArticle"
        imported:
          type: integer
        total:
          type: integer
      required:
        - articles
        - imported
        - total
    ScrapedArticle:
      type: object
      properties:
        title:
          type: string
        originalTitle:
          type: string
        excerpt:
          type: string
        content:
          type: string
        originalUrl:
          type: string
        sourceName:
          type: string
        coverImage:
          type: string
          nullable: true
        publishedAt:
          type: string
          nullable: true
        imported:
          type: boolean
      required:
        - title
        - originalTitle
        - excerpt
        - content
        - originalUrl
        - sourceName
        - imported
    TranslateRequest:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        excerpt:
          type: string
        sourceName:
          type: string
        sourceLanguage:
          type: string
          default: "es"
      required:
        - title
        - content
        - excerpt
        - sourceName
    TranslateResponse:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        excerpt:
          type: string
      required:
        - title
        - content
        - excerpt
...