# lib/api-zod/src/generated/types/fetchSourceRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * La Liga Brasil API
 * OpenAPI spec version: 0.1.0
 */

export interface FetchSourceRequest {
  sourceId: string;
  maxArticles?: number;
  autoImport?: boolean;
}