# lib/api-zod/src/generated/types/fetchSourceResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * La Liga Brasil API
 * OpenAPI spec version: 0.1.0
 */
import type { ScrapedArticle } from "./scrapedArticle";

export interface FetchSourceResponse {
  articles: ScrapedArticle[];
  imported: number;
  total: number;
}