# lib/api-zod/src/generated/types/team.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * La Liga Brasil API
 * OpenAPI spec version: 0.1.0
 */

export interface Team {
  id: number;
  name: string;
  slug: string;
  shortName: string;
  city: string;
  stadium: string;
  foundedYear?: number | null;
  description?: string | null;
  logoUrl?: string | null;
  primaryColor: string;
  secondaryColor: string;
  articleCount: number;
}