# lib/api-zod/src/generated/types/translateRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * La Liga Brasil API
 * OpenAPI spec version: 0.1.0
 */

export interface TranslateRequest {
  title: string;
  content: string;
  excerpt: string;
  sourceName: string;
  sourceLanguage?: string;
}